refactoring: toast component  - component  - docs
