--- src/components/Toast/components/ToastProvider/ToastProvider.css
@@ -1,49 +1,0 @@
-
-.slide-left-enter-active,
-.slide-left-leave-active {
-    transition: all 0.5s ease;
-}
-.slide-left-enter-from,
-.slide-left-leave-to {
-    opacity: 0;
-    transform: translateX(30px);
-}
-
-.slide-right-enter-active,
-.slide-right-leave-active {
-    transition: all 0.5s ease;
-}
-.slide-right-enter-from,
-.slide-right-leave-to {
-    opacity: 0;
-    transform: translateX(-30px);
-}
-
-.slide-top-enter-active,
-.slide-top-leave-active {
-    transition: all 0.5s ease;
-}
-.slide-top-enter-from,
-.slide-top-leave-to {
-    opacity: 0;
-    transform: translateY(30px);
-}
-
-.slide-bottom-enter-active,
-.slide-bottom-leave-active {
-    transition: all 0.5s ease;
-}
-.slide-bottom-enter-from,
-.slide-bottom-leave-to {
-    opacity: 0;
-    transform: translateY(-30px);
-}
-
-.fade-enter-active,
-.fade-leave-active {
-    transition: all 0.5s ease;
-}
-.fade-enter-from,
-.fade-leave-to {
-    opacity: 0;
-}
