--- src/components/Toast/components/ToastProvider/types.ts
@@ -1,22 +1,0 @@
-import type { ToastType } from '@/components/Toast/types'
-import type { DefineComponent } from 'vue'
-
-export type ToastItem = {
-    time: number // ms
-    type: ToastType
-    text: string
-    component?: DefineComponent
-    componentProps?: Record<string, unknown>
-}
-
-export type ToastItemWithId = ToastItem & {
-    id: string
-}
-
-export type ToastTransition = 'slide-left' | 'slide-right' | 'fade' | 'slide-top' | 'slide-bottom'
-
-export type UseToastInjection = {
-    add: (toast: ToastItem) => string
-    remove: (id: string) => boolean // true if removed, false if not found
-    pop: () => string // empty '' string if no toast to pop
-}
