--- src/components/Toast/types.ts
@@ -1,2 +1,0 @@
-export type ToastType = 'success' | 'warning' | 'danger' | 'empty'
-export type ToastAlign = 'start' | 'center' | 'end'
